requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.6=pypi_0
aiohttp=3.11.12=pypi_0
aiosignal=1.3.2=pypi_0
alluka=0.3.3=pypi_0
asyncpg=0.30.0=pypi_0
attrs=24.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.31=h06a4308_0
colorlog=6.9.0=pypi_0
expat=2.6.4=h6a678d5_0
frozenlist=1.5.0=pypi_0
hikari=2.0.0=pypi_0
hikari-lightbulb=2.3.5=pypi_0
hikari-miru=4.2.0=pypi_0
idna=3.10=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
openssl=3.0.15=h5eee18b_0
pip=25.0=py312h06a4308_0
propcache=0.2.1=pypi_0
python=3.12.9=h5148396_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
setuptools=75.8.0=py312h06a4308_0
sqlite=3.45.3=h5eee18b_0
tk=8.6.14=h39e8969_0
typing-extensions=4.12.2=pypi_0
tzdata=2025a=h04d1e81_0
wheel=0.45.1=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
zlib=1.2.13=h5eee18b_1
